feat: add husky lint-staged

+++ contracts/IWETH.sol
@@ -4,5 +4,6 @@ pragma solidity ^0.8.0;
 
 interface IWETH {
     function deposit() external payable;
-	function approve(address spender, uint256 amount) external returns (bool);
+
+    function approve(address spender, uint amount) external returns (bool);
 }
